Cleanup Payload objects
